+++ projects/ngx-scroll-to/src/lib/scroll-to-animation.ts
@@ -1,99 +1,96 @@
 import { Observable, ReplaySubject } from 'rxjs';
 
 import { EASING } from './scroll-to-helpers';
-import {
-  ScrollToConfigOptions,
-  ScrollToListenerTarget
-} from './scroll-to-config.interface';
+import { ScrollToConfigOptions, ScrollToListenerTarget } from './scroll-to-config.interface';
 
 /** Scroll To Animation */
 export class ScrollToAnimation {
 
   /** Number of milliseconds for each Tick */
-  private _tick: number;
+  private tick: number;
 
   /** Interval */
-  private _interval: any;
+  private interval: any;
 
   /** Time Lapsed in milliseconds */
-  private _timeLapsed: number;
+  private timeLapsed: number;
 
   /** Percentage of time lapsed */
-  private _percentage: number;
+  private percentage: number;
 
   /** Position of the Element */
-  private _position: number;
+  private position: number;
 
   /** Last Element Position */
-  private _lastPosition: number;
+  private lastPosition: number;
 
   /** Start Position of the Element */
-  private _startPosition: number;
+  private startPosition: number;
 
   /** The Distance to scroll */
-  private _distance: number;
+  private distance: number;
 
   /** Observable Source */
-  private _source$: ReplaySubject<number>;
+  private source$: ReplaySubject<number>;
 
   /** Scroll Top of the Window */
-  private _windowScrollTop: number;
+  private windowScrollTop: number;
 
   /** Mapped Offset taken from the active Offset Map */
-  private _mappedOffset: number;
+  private mappedOffset: number;
 
   /**
    * Class Constructor.
    *
-   * @param _container            The Container
-   * @param _listenerTarget       The Element that listens for DOM Events
-   * @param _isWindow             Whether or not the listener is the Window
-   * @param _to                   Position to scroll to
-   * @param _options              Additional options for scrolling
-   * @param _isBrowser            Whether or not execution runs in the browser
+   * @param container            The Container
+   * @param listenerTarget       The Element that listens for DOM Events
+   * @param isWindow             Whether or not the listener is the Window
+   * @param to                   Position to scroll to
+   * @param options              Additional options for scrolling
+   * @param isBrowser            Whether or not execution runs in the browser
    *                              (as opposed to the server)
    */
   constructor(
-    private _container: HTMLElement,
-    private _listenerTarget: ScrollToListenerTarget,
-    private readonly _isWindow: boolean,
-    private readonly _to: number,
-    private readonly _options: ScrollToConfigOptions,
-    private _isBrowser: boolean
+    private container: HTMLElement,
+    private listenerTarget: ScrollToListenerTarget,
+    private readonly isWindow: boolean,
+    private readonly to: number,
+    private readonly options: ScrollToConfigOptions,
+    private isBrowser: boolean
   ) {
-    this._tick = 16;
-    this._interval = null;
-    this._lastPosition = null;
-    this._timeLapsed = 0;
+    this.tick = 16;
+    this.interval = null;
+    this.lastPosition = null;
+    this.timeLapsed = 0;
 
-    this._windowScrollTop = window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop || 0;
+    this.windowScrollTop = window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop || 0;
 
-    if (!this._container) {
-      this._startPosition = this._windowScrollTop;
+    if (!this.container) {
+      this.startPosition = this.windowScrollTop;
     } else {
-      this._startPosition = this._isWindow ? this._windowScrollTop : this._container.scrollTop;
+      this.startPosition = this.isWindow ? this.windowScrollTop : this.container.scrollTop;
     }
 
     // Correction for Starting Position of nested HTML Elements
-    if (this._container && !this._isWindow) {
-      this._to = this._to - this._container.getBoundingClientRect().top + this._startPosition;
+    if (this.container && !this.isWindow) {
+      this.to = this.to - this.container.getBoundingClientRect().top + this.startPosition;
     }
 
     // Set Distance
-    const directionalDistance = this._startPosition - this._to;
-    this._distance = this._container ? Math.abs(this._startPosition - this._to) : this._to;
+    const directionalDistance = this.startPosition - this.to;
+    this.distance = this.container ? Math.abs(this.startPosition - this.to) : this.to;
 
-    this._mappedOffset = this._options.offset;
+    this.mappedOffset = this.options.offset;
 
     // Set offset from Offset Map
-    if (this._isBrowser) {
-      this._options
+    if (this.isBrowser) {
+      this.options
         .offsetMap
-        .forEach((value, key) => this._mappedOffset = window.innerWidth > key ? value : this._mappedOffset);
+        .forEach((value, key) => this.mappedOffset = window.innerWidth > key ? value : this.mappedOffset);
     }
 
-    this._distance += this._mappedOffset * (directionalDistance <= 0 ? 1 : -1);
-    this._source$ = new ReplaySubject();
+    this.distance += this.mappedOffset * (directionalDistance <= 0 ? 1 : -1);
+    this.source$ = new ReplaySubject();
   }
 
   /**
@@ -101,45 +98,45 @@ export class ScrollToAnimation {
    *
    * @returns         Observable containing a number
    */
-  public start(): Observable<number> {
-    clearInterval(this._interval);
-    this._interval = setInterval(this._loop, this._tick);
-    return this._source$.asObservable();
-  }
-
-  /** Recursively loop over the Scroll Animation */
-  private _loop = (): void => {
-
-    this._timeLapsed += this._tick;
-    this._percentage = (this._timeLapsed / this._options.duration);
-    this._percentage = (this._percentage > 1) ? 1 : this._percentage;
-
-    // Position Update
-    this._position = this._startPosition +
-      ((this._startPosition - this._to <= 0 ? 1 : -1) *
-        this._distance *
-        EASING[this._options.easing](this._percentage));
-
-    if (this._lastPosition !== null && this._position === this._lastPosition) {
-      this.stop();
-    } else {
-      this._source$.next(this._position);
-      this._isWindow
-        ? this._listenerTarget.scrollTo(0, Math.floor(this._position))
-        : this._container.scrollTop = Math.floor(this._position);
-      this._lastPosition = this._position;
-    }
+  start(): Observable<number> {
+    clearInterval(this.interval);
+    this.interval = setInterval(this.loop, this.tick);
+    return this.source$.asObservable();
   }
 
   /**
    * Stop the current Scroll Animation Loop.
    *
-   * @param force 			    Force to stop the Animation Loop
+   * @param force          Force to stop the Animation Loop
    * @returns               Void
    */
-  public stop(): void {
-    clearInterval(this._interval);
-    this._interval = null;
-    this._source$.complete();
+  stop(): void {
+    clearInterval(this.interval);
+    this.interval = null;
+    this.source$.complete();
+  }
+
+  /** Recursively loop over the Scroll Animation */
+  private loop = (): void => {
+
+    this.timeLapsed += this.tick;
+    this.percentage = (this.timeLapsed / this.options.duration);
+    this.percentage = (this.percentage > 1) ? 1 : this.percentage;
+
+    // Position Update
+    this.position = this.startPosition +
+      ((this.startPosition - this.to <= 0 ? 1 : -1) *
+        this.distance *
+        EASING[this.options.easing](this.percentage));
+
+    if (this.lastPosition !== null && this.position === this.lastPosition) {
+      this.stop();
+    } else {
+      this.source$.next(this.position);
+      this.isWindow
+        ? this.listenerTarget.scrollTo(0, Math.floor(this.position))
+        : this.container.scrollTop = Math.floor(this.position);
+      this.lastPosition = this.position;
+    }
   }
 }