+++ projects/ngx-scroll-to/src/public_api.ts
@@ -1,5 +1,0 @@
-export { ScrollToConfigOptions, ScrollToOffsetMap } from './lib/scroll-to-config.interface';
-export { ScrollToAnimationEasing } from './lib/scroll-to-easing.interface';
-export { ScrollToDirective } from './lib/scroll-to.directive';
-export { ScrollToModule } from './lib/scroll-to.module';
-export { ScrollToService } from './lib/scroll-to.service';